requirements.txt
click==8.1.3
Flask==2.1.1
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
uvicorn==0.22.0
waitress==2.1.2
Werkzeug==2.3.4
telepota